requirements.txt
psutil==5.9.8
requests==2.31.0
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
